Adding some performance improvements
Specially on the huge memory consumption

+++ src/animatedlayer.h
@@ -10,15 +10,15 @@ class AnimatedLayer : public Layer
 {
     Q_OBJECT
 
-    Q_PROPERTY(int horizontalStep READ horizontalStep WRITE setHorizontalStep NOTIFY horizontalStepChanged)
+    Q_PROPERTY(qreal horizontalStep READ horizontalStep WRITE setHorizontalStep NOTIFY horizontalStepChanged)
     Q_PROPERTY(Quasi::LayerDirection direction READ direction WRITE setDirection NOTIFY directionChanged)
 
 public:
     AnimatedLayer(Layer *parent = 0);
     ~AnimatedLayer();
 
-    int horizontalStep() const { return m_horizontalStep; };
-    void setHorizontalStep(const int &step);
+    qreal horizontalStep() const { return m_horizontalStep; };
+    void setHorizontalStep(const qreal &step);
 
     Quasi::LayerDirection direction() const { return m_direction; };
     void setDirection(const Quasi::LayerDirection &direction);
@@ -32,8 +32,8 @@ signals:
 private:
     void updateHorizontalStep();
 
-    int m_horizontalStep;
-    int m_currentHorizontalStep;
+    qreal m_horizontalStep;
+    qreal m_currentHorizontalStep;
     Quasi::LayerDirection m_direction;
 };
